requirements.txt
httpx>=0.27
fastapi>=0.115
uvicorn[standard]>=0.32
jinja2>=3.1
python-multipart>=0.0.9
apscheduler>=3.10
feedgen>=1.0
pillow>=10.0
pydantic>=2.0
pydantic-settings>=2.0
aiofiles>=24.1

# Testing
pytest>=8.0
anyio>=4.0
pytest-anyio>=0.0.0
